requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
jinja2>=3.0.0
pytest>=7.0.0
python-dateutil>=2.8.2
markdown>=3.4.0
pdfkit>=1.0.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6